Move `serialize` into the scrypt module
