+++ src/internal.js
@@ -5,6 +5,5 @@
 // In the future we hope to minimize / reduce this
 
 export { hmacSha256 } from './util/crypto/crypto.js'
-export * from './util/decorators.js'
 export * from './util/encoding.js'
 export * from './util/util.js'
